refactor(automations): move slackWebhookGuard to automations-server clients/slack (ADR-063 P1)
First delivery-tier move, unblocked by the DispatchError extraction — the
guard's only dependency. Both importers rewired to the package path.
