+++ langwatch/src/server/app-layer/automations/delivery/triggerNotifier.ts
@@ -5,7 +5,7 @@ import {
 import type { TriggerNotifier } from "~/server/app-layer/automations/trigger-template.service";
 import { sendEmail } from "~/server/mailer/emailSender";
 import { assertWebhookDelivered, sendWebhook } from "./sendWebhook";
-import { isSlackWebhookUrl } from "./slackWebhookGuard";
+import { isSlackWebhookUrl } from "@langwatch/automations-server/clients/slack/webhook-guard";
 import { postSlackChatMessage } from "./slackWebApi";
 
 /**
